Move container resource cleanup to external process

+++ IronFoundry.Warden/Containers/ContainerResourceHolder.cs
@@ -1,6 +1,10 @@
 ﻿using System.Net;
 using IronFoundry.Warden.Configuration;
 using IronFoundry.Warden.Utilities;
+using IronFoundry.Warden.Containers.Messages;
+using System.Collections.Generic;
+using System;
+using System.IO;
 
 namespace IronFoundry.Warden.Containers
 {
@@ -19,14 +23,22 @@ namespace IronFoundry.Warden.Containers
     public class ContainerResourceHolder : IResourceHolder
     {
         private const int TerminateWaitTimeout = 2000; // ms
+        private readonly bool deleteDirectories;
 
-        public ContainerResourceHolder(ContainerHandle handle, IContainerUser user, IContainerDirectory directory, JobObject jobObject, ILocalTcpPortManager localTcpPortManager)
+        public ContainerResourceHolder(
+            ContainerHandle handle, 
+            IContainerUser user, 
+            IContainerDirectory directory, 
+            JobObject jobObject, 
+            ILocalTcpPortManager localTcpPortManager, 
+            bool deleteDirectories)
         {
             Handle = handle;
             User = user;
             Directory = directory;
             JobObject = jobObject;
             LocalTcpPortManager = localTcpPortManager;
+            this.deleteDirectories = deleteDirectories;
         }
 
         public ushort? AssignedPort { get; set; }
@@ -41,42 +53,67 @@ namespace IronFoundry.Warden.Containers
             JobObject.TerminateProcessesAndWait(TerminateWaitTimeout);
             JobObject.Dispose();
 
-            Directory.Delete();
-            User.Delete();
+            try
+            {
+                if (deleteDirectories)
+                    Directory.Delete();
+            }
+            catch (IOException)
+            {
+            }
+
             if (AssignedPort.HasValue)
             {
-                LocalTcpPortManager.ReleaseLocalPort(AssignedPort.Value, User.UserName);
+                try
+                {
+                    LocalTcpPortManager.ReleaseLocalPort(AssignedPort.Value, User.UserName);
+                }
+                catch (Exception)
+                {
+                    // TODO: Log
+                }
+            }
+
+            try
+            {
+                User.Delete();
+            }
+            catch (System.Exception)
+            {
+                // TODO: Add logging for cleanup case
             }
         }
 
         public static IResourceHolder Create(IWardenConfig config)
-        {
+        {   
             var handle = new ContainerHandle();
             var user = ContainerUser.CreateUser(handle, new LocalPrincipalManager(new DesktopPermissionManager()));
-            var directory = new ContainerDirectory(handle, user, true, config);
+            var directory = new ContainerDirectory(handle, user, config.ContainerBasePath, true);
             var localPortManager = new LocalTcpPortManager(new FirewallManager(), new NetShRunner());
             var resoureHolder = new ContainerResourceHolder(
                 handle,
                 user,
                 directory,
                 new JobObject(handle.ToString()),
-                localPortManager
+                localPortManager,
+                config.DeleteContainerDirectories
                 );
 
             return resoureHolder;
         }
 
-        public static IResourceHolder Create(IWardenConfig config, ContainerHandle handle)
+        public static IResourceHolder CreateForDestroy(IWardenConfig config, ContainerHandle handle)
         {
             var user = new TempUser(handle, new LocalPrincipalManager(new DesktopPermissionManager()));
-            var directory = new ContainerDirectory(handle, user, true, config);
+            var directory = new TempDirectory(handle, config.ContainerBasePath);
             var localPortManager = new LocalTcpPortManager(new FirewallManager(), new NetShRunner());
             var resoureHolder = new ContainerResourceHolder(
                 handle,
                 user,
                 directory,
                 new JobObject(handle.ToString()),
-                localPortManager
+                localPortManager,
+                config.DeleteContainerDirectories
                 );
 
             return resoureHolder;
@@ -102,6 +139,37 @@ namespace IronFoundry.Warden.Containers
             public void Delete()
             {
                 userManager.DeleteUser(UserName);
+            }
+        }
+
+        /// <summary>
+        /// Exists to create a temporary implementation primarily for deleting the directory.
+        /// </summary>
+        class TempDirectory : IContainerDirectory
+        {
+            private readonly string fullPath;
+
+            public TempDirectory(ContainerHandle handle, string containerBasePath)
+            {
+                fullPath = Path.Combine(containerBasePath, handle);
+            }
+
+            public string FullName
+            {
+                get { return fullPath; }
+            }
+
+            public void BindMounts(IEnumerable<BindMount> mounts)
+            {
+                throw new NotImplementedException();
+            }
+
+            public void Delete()
+            {
+                if (System.IO.Directory.Exists(fullPath))
+                {
+                    System.IO.Directory.Delete(fullPath, true);
+                }
             }
         }
     }
